+++ vojuck.victoria/P4/inputMatrix.cpp
@@ -1,7 +1,7 @@
 #include "inputMatrix.hpp"
 #include <algorithm>
 
-size_t inputArray(std::istream & in, int * array, size_t size, size_t toread)
+size_t inputArray(std::istream & in, long long int * array, size_t size, size_t toread)
 {
   for (size_t i = 0; i < std::min(size, toread); ++i)
   {
