+++ vojuck.victoria/P4/isUpperTriangular.hpp
@@ -2,6 +2,6 @@
 #define ISUPPERTRIANGULAR_HPP
 #include <cstddef>
 
-bool isUpperTriangular(int* matrix, size_t rows, size_t cols);
+bool isUpperTriangular(long long int* matrix, size_t rows, size_t cols);
 
 #endif
